Opdatering af application.properties til at vise detaljeret fejlkode. derudover rettelse af constructor i wishlist, som krævede int i stedet for long. derudover rettelse af headernavn i homepage samt ændring af "/" til at henvise til homepage

+++ src/main/java/org/example/projekt4_gruppe_3/Model/Wishlist.java
@@ -39,7 +39,7 @@ public class Wishlist {
     public Wishlist() {
     }
 
-    public Wishlist(String listName, String listDescription, int lastUpdated, String listImg) {
+    public Wishlist(String listName, String listDescription, Long lastUpdated, String listImg) {
         this.name = listName;
         this.description = listDescription;
         this.lastUpdated = lastUpdated;
